Refactoring Route into an interface and an OutdoorRoute For future indoor-route usage and mapping, it seemed to be beneficial to abstract routing into an additional layer. No impact on test coverage.
