+++ src/app/core/models/index.ts
@@ -4,7 +4,7 @@ export * from './campus';
 export * from './coordinates';
 export * from './outdoor-poi';
 export * from './poi';
-export * from './route';
+export * from './outdoor-route';
 export * from './route-step';
 export * from './transport-mode';
 export * from './map';
