+++ src/app/routes/components/routes-list/routes-list.component.spec.ts
@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 import { RoutesListComponent } from './routes-list.component';
 import { RouterTestingModule } from '@angular/router/testing';
-import { Route } from 'src/app/core';
+import { OutdoorRoute } from 'src/app/core';
 
 describe('RoutesListComponent', () => {
     let component: RoutesListComponent;
