+++ src/app/shared/state.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
-import { Route } from '../core/models';
+import { OutdoorRoute } from '../core/models';
 
 @Injectable({
   providedIn: 'root'
 })
 export class StateService {
-  sharedRoute: Route;
+  sharedRoute: OutdoorRoute;
 
   constructor() { }
 }
